some fixes

+++ Backend/tickets/src/main/java/com/event/tickets/domain/CreateTicketTypeRequest.java
@@ -1,16 +1,17 @@
 package com.event.tickets.domain;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/** FIX #15: price changed from Double to BigDecimal for financial precision. */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class CreateTicketTypeRequest {
-
-  private String name;
-  private Double price;
-  private String description;
-  private Integer totalAvailable;
+    private String name;
+    private BigDecimal price;
+    private String description;
+    private Integer totalAvailable;
 }
