+++ Backend/tickets/src/main/java/com/event/tickets/domain/UpdateTicketTypeRequest.java
@@ -1,18 +1,19 @@
 package com.event.tickets.domain;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/** FIX #15: price changed from Double to BigDecimal for financial precision. */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class UpdateTicketTypeRequest {
-
-  private UUID id;
-  private String name;
-  private Double price;
-  private String description;
-  private Integer totalAvailable;
+    private UUID id;
+    private String name;
+    private BigDecimal price;
+    private String description;
+    private Integer totalAvailable;
 }
