+++ Backend/tickets/src/main/java/com/event/tickets/domain/dtos/ListEventTicketTypeResponseDto.java
@@ -1,19 +1,19 @@
 package com.event.tickets.domain.dtos;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/** FIX #15: price changed from Double to BigDecimal. */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ListEventTicketTypeResponseDto {
-
-  private UUID id;
-  private String name;
-  private Double price;
-  private String description;
-  private Integer totalAvailable;
-
+    private UUID id;
+    private String name;
+    private BigDecimal price;
+    private String description;
+    private Integer totalAvailable;
 }
